requirements.txt
fastapi
uvicorn
fastmcp
httpx
pydantic
pydantic-settings
regex
schedule
json-repair
openai
python-dotenv
starlette
typing-extensions